refactor: split public type definitions
