+++ src/aggregate-id.ts
@@ -1,0 +1,7 @@
+interface AggregateId {
+  typeName: string;
+  value: string;
+  asString: () => string;
+}
+
+export type { AggregateId };
